+++ apps/builder/app/builder/features/style-panel/shared/css-editor/copy-paste-menu.tsx
@@ -13,21 +13,20 @@ import {
   toValue,
   type StyleMap,
 } from "@webstudio-is/css-engine";
-import { useStore } from "@nanostores/react";
-import { $advancedStylesLonghands } from "./stores";
 
 export const copyAttribute = "data-declaration";
 
 export const CopyPasteMenu = ({
   children,
   properties,
+  styleMap,
   onPaste,
 }: {
   children: ReactNode;
   properties: Array<string>;
+  styleMap: StyleMap;
   onPaste: (cssText: string) => void;
 }) => {
-  const advancedStylesLonghands = useStore($advancedStylesLonghands);
   const lastClickedProperty = useRef<string>();
 
   const handlePaste = () => {
@@ -38,7 +37,7 @@ export const CopyPasteMenu = ({
     // We want to only copy properties that are currently in front of the user.
     // That includes search or any future filters.
     const currentStyleMap: StyleMap = new Map();
-    for (const [property, value] of advancedStylesLonghands) {
+    for (const [property, value] of styleMap) {
       const isEmpty = toValue(value) === "";
       if (properties.includes(property) && isEmpty === false) {
         currentStyleMap.set(hyphenateProperty(property), value);
@@ -55,7 +54,7 @@ export const CopyPasteMenu = ({
     if (property === undefined) {
       return;
     }
-    const value = advancedStylesLonghands.get(property);
+    const value = styleMap.get(property);
 
     if (value === undefined) {
       return;
