+++ apps/builder/app/builder/features/style-panel/shared/css-editor/css-editor.stories.tsx
@@ -1,6 +1,6 @@
 import { getStyleDeclKey, type StyleDecl } from "@webstudio-is/sdk";
 import { registerContainers } from "~/shared/sync";
-import { Section } from "./advanced";
+import { CssEditor as CssEditorComponent } from "./css-editor";
 import {
   $breakpoints,
   $instances,
@@ -8,7 +8,7 @@ import {
   $styles,
   $styleSourceSelections,
 } from "~/shared/nano-states";
-import { setProperty } from "../../shared/use-style-data";
+import { setProperty } from "../use-style-data";
 import { $awareness } from "~/shared/awareness";
 
 const backgroundImage: StyleDecl = {
@@ -42,11 +42,24 @@ setProperty("accentColor")({ type: "keyword", value: "red" });
 setProperty("alignContent")({ type: "keyword", value: "normal" });
 setProperty("opacity")({ type: "unit", unit: "number", value: 11.2 });
 
-export const Advanced = () => {
-  return <Section />;
+export const CssEditor = () => {
+  const styleMap = new Map([
+    ["backgroundImage", backgroundImage.value],
+    ["accentColor", { type: "keyword", value: "red" }],
+    ["alignContent", { type: "keyword", value: "normal" }],
+    ["opacity", { type: "unit", unit: "number", value: 11.2 }],
+  ]);
+  return (
+    <CssEditorComponent
+      styleMap={styleMap}
+      deleteProperty={() => undefined}
+      setProperty={() => () => undefined}
+      insertProperties={() => undefined}
+    />
+  );
 };
 
 export default {
-  title: "Style Panel/Advanced",
-  component: Advanced,
+  title: "Style Panel",
+  component: CssEditor,
 };
