+++ apps/builder/app/builder/features/style-panel/shared/css-editor/index.ts
@@ -1,0 +1,1 @@
+export { CssEditor, type CssEditorApi } from "./css-editor";
